added home page and own components for drinkdetailles that can be use from both us
